requirements.txt
appnope==0.1.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.154
botocore==1.12.154
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
docutils==0.14
enum34==1.1.6
futures==3.2.0
gensim==0.12.4
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
jmespath==0.9.4
nltk==3.4.1
numpy==1.11.0
oauthlib==3.0.1
pandas==0.18.1
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
pprint==0.1
prompt-toolkit==1.0.16
ptyprocess==0.6.0
pyenchant==2.0.0
Pygments==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
python-rake==1.4.5
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
scandir==1.10.0
scikit-learn==0.20.3
scipy==1.2.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.3
soupsieve==1.9.1
traitlets==4.3.2
tweepy==3.7.0
urllib3==1.25.2
vaderSentiment==3.2.1
wcwidth==0.1.7
wikipedia==1.4.0
